move blink_manager to gpui
