+++ crates/gpui/src/input.rs
@@ -3,10 +3,12 @@ mod bidi;
 ///
 /// Explicitly not exported using `pub use bindings::*` to avoid namespace pollution.
 pub mod bindings;
+mod blink_manager;
 mod handler;
 mod state;
 
 pub use bidi::{TextDirection, detect_base_direction};
 pub use bindings::{INPUT_CONTEXT, InputBindings, bind_input_keys};
+pub use blink_manager::BlinkManager;
 pub use handler::*;
 pub use state::{InputLineLayout, InputState};
